+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Add `resolveInvoiceRuleP` and `resolveInvoiceRule` methods that return an invoice rule for the given brand, model and country - [ripe-pulse/#291](https://github.com/ripe-tech/ripe-pulse/issues/291)
+* Added `getShipments/getShipmentsP` and `createShipment/createShipmentsP` - [ripe-pulse/#276](https://github.com/ripe-tech/ripe-pulse/issues/276)
 
 ### Changed